Resolve merge conflicts, include loading file messages
